Extract shared Duckgres helpers into ducklake types
